+++ ui/main/game/galactic_war/cards/gwaio_upgrade_laserdefensetower.js
@@ -4,7 +4,7 @@ define([
   return {
     visible: _.constant(true),
     describe: _.constant(
-      "!LOC:Laser Defense Tower Upgrade Tech increases the turn speed of the turret by 200%."
+      "!LOC:Laser Defense Tower Upgrade Tech increases the range of the turret by 25%."
     ),
     summarize: _.constant("!LOC:Laser Defense Tower Upgrade Tech"),
     icon: _.constant(
@@ -49,9 +49,9 @@ define([
       var mods = [
         {
           file: "/pa/units/land/laser_defense/laser_defense_tool_weapon.json",
-          path: "yaw_rate",
+          path: "max_range",
           op: "multiply",
-          value: 3,
+          value: 1.25,
         },
       ];
       inventory.addMods(mods);