+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Simplified upgrade tech checks to improve future compatibility
 - The following unit upgrades:
   - Angel
+  - Laser Defense Tower
 
 ### Fixed